Add long conversion

+++ train_test/train_test_TSVIT_PASTIS.py
@@ -26,7 +26,7 @@ def train_model(model, train_loader, criterion, optimizer, num_epochs=50):
     for epoch in range(num_epochs):
         running_loss = 0.0
         for inputs, labels, time_points in train_loader:
-            labels = labels.squeeze(-1)
+            labels = labels.squeeze(-1).long()
             inputs = inputs.permute(0, 1, 3, 4, 2)  # New shape: [8, 37, 128, 128, 10]
             inputs, labels = inputs.to(device), labels.to(device)
             print(f"Labels shape: {labels.shape}")
